+++ e2e_test/lib/test_stringCtor.dart
@@ -1,0 +1,3 @@
+String createAString() {
+  return new String.fromCharCode(10);
+}