+++ dart2ts/CHANGELOG.md
@@ -2,6 +2,8 @@
  - fixed issue with multiple fields init
  - better handling of static methods on native types
  - better ordering of emitted declaration
+ - missing not operator
+ - better handling of constructors for native types (string)
 
 ### v0.9.13
  - fixed string escape nasty bug